+++ tests/test_models.py
@@ -1812,6 +1812,7 @@ TRANSLATABLE_LABEL_MODELS = [
     oauth2_models.AbstractRefreshToken,
     oauth2_models.AbstractIDToken,
     oauth2_models.AbstractDeviceGrant,
+    oauth2_models.AbstractPushedAuthorizationRequest,
 ]